+++ xenon-common/src/main/java/com/vmware/xenon/services/common/QueryTaskFactoryService.java
@@ -24,6 +24,7 @@ public class QueryTaskFactoryService extends FactoryService {
 
     public QueryTaskFactoryService() {
         super(QueryTask.class);
+        super.setPeerNodeSelectorPath(ServiceUriPaths.DEFAULT_1X_NODE_SELECTOR);
     }
 
     @Override